Relax the atomic heap interface
To support an implementation with per-location locks, the heap interface had to
be extended with an invariant that is chosen by the implementation and required
by all operations. I also added `CombineSepGives` and `CombineSepAs`
instances (mostly) for convenience.

The implementation uses per-location locks. The interface is extended with an
invariant passed as hypothesis to all operations.